Fragmentshader clean up. Removal of several uniforms (5) and replacement by two other uniforms instead. Refactoring/Optimization of main.cpp. Some variables are now only instantiated once instead of in each while loop. Translation of Kurt Spencer's Java version of OpenSimplex Noise to C++. Extension of Water class to use OpenSimplex Noise to make waves appear more random. Adding new water textures. General heightmap change. Vertex x and z coordinates are now always in interval [0,1] instead of an arbitrary value. Camera settings changed because of heightmap improvements. Camera speed can now easily be adjusted since the heightmap is always the same scale. RMP height values had to be adjusted because of heightmap changes. Water wave settings also had to be adjusted. Extended ThermalErosion algorithm to consider material properties. E.g. rock is hard, sand is soft.

+++ TerrainGeneration/TerrainGeneration/Heightmap.hpp
@@ -33,7 +33,7 @@ public:
     void setHeightAt(int column, int row, float height);
     float getHeightAt(int column, int row);
     float getSlopeAt(int column, int row);
-    void loadHeightmap(string filename, int strength);
+    void loadHeightmap(string filename, float strength);
     float getMinHeight();
     float getMaxHeight();
 };
